Replace requests.get(…).json() with json.loads(…)
In the server environment on w3c-test.org, calling .json() on the result of
from a requests.get(…) fails with a `'list' object is not callable` error.
I think it might be a bug in the version of the Python "requests" module we
have on w3c-test.org but I don't really know. But calling .text on the
result works fine, so we can just use json.loads(requests.get(…).text)
instead and achieve the same effect. So let's just do that.